First very very beta homepage for smc app
t has a bug that causes it to not emulate
when found and removed it says failed to open on the device, but here
are a few things that I’ve put together for the home screen. I was
thinking for color scheme we can make it kind of like the way google
now and the play store look like asthetically.